# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Order Fulfillment and Logistics Hub API
  version: 0.1.0
  description: >-
    Contract for BUSINESS administration, trusted internal intake, and isolated
    DemoScenario operations. JSON field names use snake_case; identifiers and
    NUMERIC(18,3) quantities are strings.
servers:
  - url: http://localhost:8080
    description: Local Docker Compose
...
security:
  - basicAuth: []
paths:
...
  /api/v1/orders/{order_id}:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    get:
      tags: [Orders]
      operationId: getOrder
      responses:
        '200':
          description: Aggregated BUSINESS order detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderDetail' }
        default: { $ref: '#/components/responses/ErrorResponse' }
...
  /api/v1/review-cases:
    get:
      tags: [Reviews]
      operationId: listReviewCases
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: status
          in: query
          schema: { type: string, enum: [OPEN, RESOLVED, DISMISSED] }
        - name: reason_code
          in: query
          schema: { type: string, maxLength: 64 }
        - name: responsible_team
          in: query
          schema: { type: string, maxLength: 128 }
        - name: source_channel
          in: query
          description: Limits the queue to BUSINESS orders from this source channel.
          schema: { $ref: '#/components/schemas/SourceChannel' }
      responses:
        '200':
          description: BUSINESS manual-review queue
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewCasePage' }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/review-cases/{case_id}:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    get:
      tags: [Reviews]
      operationId: getReviewCase
      responses:
        '200':
          description: Review subject, evidence, suggestions, and allowed actions
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewCase' }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/review-cases/{case_id}/resolve-customer:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    post:
      tags: [Reviews]
      operationId: resolveCustomerReview
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Operator'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResolveCustomerReviewCommand' }
      responses:
        '200':
          description: Existing active BUSINESS customer linked and review resolved atomically; no customer is created by this command
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewCase' }
        '409':
          description: expected_version is stale, the review is no longer OPEN, or evidence/mapping conflicts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/review-cases/{case_id}/resolve-sku:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    post:
      tags: [Reviews]
      operationId: resolveSkuReview
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Operator'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResolveSkuReviewCommand' }
      responses:
        '200':
          description: Existing active SKU and explicit source mapping saved atomically; no SKU is created by this command
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewCase' }
        '409':
          description: expected_version is stale, the review is no longer OPEN, or evidence/mapping conflicts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        default: { $ref: '#/components/responses/ErrorResponse' }
...
  /api/v1/review-cases/{case_id}/resolve-jd-tracking-conflict:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    post:
      tags: [Reviews]
      operationId: resolveJdTrackingConflict
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Operator'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VersionedNoteCommand' }
      responses:
        '200':
          description: Human review resolved the blocking JD tracking, Carrier-mapping, or terminal-exception conflict without creating or changing Tracking facts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewCase' }
        '409':
          description: expected_version is stale, the review is no longer OPEN, or its shipment evidence does not match the business order
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/review-cases/{case_id}/complete-source-followup:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    post:
      tags: [Reviews]
      operationId: completeSourceFollowup
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Operator'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompleteSourceFollowupCommand' }
      responses:
        '200':
          description: Review, processing stage, event, version, and audit committed atomically
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewCase' }
        '409':
          description: expected_version is stale or terminal-fulfillment/all-real-shipment-tracking gates are not satisfied
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/review-cases/{case_id}/resolve:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    post:
      tags: [Reviews]
      operationId: resolveReviewManually
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Operator'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VersionedNoteCommand' }
      responses:
        '200':
          description: Whitelisted review (master-data conflict, revision-after-export, quantity scale, fulfillment exception, sync failure, stock block, message chain needs-review, ...) closed as manually resolved with optional note
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewCase' }
        '409':
          description: expected_version is stale, the review is no longer OPEN, or the reason code is not in the manual-resolution whitelist
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/review-cases/{case_id}/dismiss:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    post:
      tags: [Reviews]
      operationId: dismissReviewCase
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Operator'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VersionedNoteCommand' }
      responses:
        '200':
          description: Open review closed as DISMISSED with optional note; message-chain reviews (drafts / reinterpret) are rejected
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReviewCase' }
        '409':
          description: expected_version is stale, the review is no longer OPEN, or message-chain reviews must be closed through their own lifecycle
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        default: { $ref: '#/components/responses/ErrorResponse' }
...
  /api/v1/procurement-tickets:
    get:
      tags: [Procurement]
      operationId: listProcurementTickets
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/ProcurementStatus' }
      responses:
        '200':
          description: Procurement workbench
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProcurementTicketPage' }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/procurement-tickets/{ticket_id}:
    parameters:
      - $ref: '#/components/parameters/TicketId'
    get:
      tags: [Procurement]
      operationId: getProcurementTicket
      responses:
        '200':
          description: Ticket items and append-only receipts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProcurementTicket' }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/procurement-tickets/{ticket_id}/retry:
    parameters:
      - $ref: '#/components/parameters/TicketId'
    post:
      tags: [Procurement]
      operationId: retryProcurement
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Operator'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VersionedNoteCommand' }
      responses:
        '201':
          description: Linked retry ticket created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProcurementTicket' }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/procurement-tickets/{ticket_id}/cancel-remaining:
    parameters:
      - $ref: '#/components/parameters/TicketId'
    post:
      tags: [Procurement]
      operationId: cancelProcurementRemaining
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Operator'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CancelRemainingCommand' }
      responses:
        '200':
          description: Remaining quantity cancelled without rolling back shipped facts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProcurementTicket' }
        default: { $ref: '#/components/responses/ErrorResponse' }
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Nginx authenticates management, internal, and demo operations; explicitly public connector callbacks use their own signatures.
  headers:
    RequestId:
      description: Request correlation identifier
      schema: { type: string }
  responses:
    ErrorResponse:
      description: Unified business or technical error
      headers:
        X-Request-Id: { $ref: '#/components/headers/RequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
...
    DateFrom:
      name: date_from
      in: query
      schema: { type: string, format: date }
    DateTo:
      name: date_to
      in: query
      schema: { type: string, format: date }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 255 }
    Operator:
      name: X-Operator
      in: header
      required: true
      description: Required at the backend seam; the public gateway overwrites this value with the authenticated principal.
      schema: { type: string, minLength: 1, maxLength: 128 }
    OrderId: { name: order_id, in: path, required: true, schema: { $ref: '#/components/schemas/Identifier' } }
...
    CaseId: { name: case_id, in: path, required: true, schema: { $ref: '#/components/schemas/Identifier' } }
...
    TicketId: { name: ticket_id, in: path, required: true, schema: { $ref: '#/components/schemas/Identifier' } }
...
  schemas:
    Identifier:
      type: string
      pattern: '^[1-9][0-9]*$'
      example: '42'
...
    SourceChannel:
      type: string
      enum: [CAISHIXIAN, JUFUBAO, FEIXIANG, WECOM]
...
    ProcurementStatus:
      type: string
      enum: [PENDING, SUCCESS, PARTIAL, FAILED, CANCELLED]
...
    PageMeta:
      type: object
      required: [page, size, total_elements, total_pages]
      properties:
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1 }
        total_elements: { type: integer, format: int64, minimum: 0 }
        total_pages: { type: integer, minimum: 0 }
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
    ApiError:
      type: object
      required: [business_code, message, http_status, request_id, trace_id]
      properties:
        business_code: { type: string, maxLength: 64 }
        message: { type: string }
        http_status: { type: integer, minimum: 400, maximum: 599 }
        request_id: { type: string }
        trace_id: { type: string }
        field_errors:
          type: array
          items: { $ref: '#/components/schemas/FieldError' }
        details:
          type: object
          additionalProperties: true
...
    Receiver:
      type: object
      required: [name, phone, address]
      properties:
        name: { type: string, maxLength: 128 }
        phone: { type: string, maxLength: 64 }
        province: { type: string, maxLength: 64 }
        city: { type: string, maxLength: 64 }
        district: { type: string, maxLength: 64 }
        town: { type: string, maxLength: 64 }
        address: { type: string, maxLength: 1000 }
    Settlement:
      type: object
      required: [method, settlement_time]
      properties:
        method: { type: string, enum: [MONTHLY, IMMEDIATE, CREDIT_TERM, PREPAID, COD, OTHER] }
        settlement_time: { type: string, format: date-time }
...
    OrderLine:
      type: object
      required: [id, line_no, line_type, product_name, specification, unit, requested_quantity, processing_stage]
      properties:
        id: { $ref: '#/components/schemas/Identifier' }
        line_no: { type: integer, minimum: 1 }
        line_type: { type: string, enum: [SINGLE, CUSTOM_BUNDLE] }
        sku_id: { $ref: '#/components/schemas/Identifier' }
        sku_code: { type: string }
        provider_id: { $ref: '#/components/schemas/Identifier' }
        product_name: { type: string }
        specification: { type: string }
        unit: { type: string }
        source_quantity: { $ref: '#/components/schemas/DecimalQuantity' }
        mapping_multiplier: { $ref: '#/components/schemas/DecimalQuantity' }
        requested_quantity: { $ref: '#/components/schemas/DecimalQuantity' }
        processing_stage: { $ref: '#/components/schemas/ProcessingStage' }
        exception_code: { type: string }
        components:
          type: array
          items: { $ref: '#/components/schemas/OrderLineComponent' }
...
    OrderSummary:
      type: object
      required: [id, order_no, source_channel, order_status, processing_stage, processing_health, completed_count, total_count, created_at, version]
      properties:
        id: { $ref: '#/components/schemas/Identifier' }
        order_no: { type: string }
        source_channel: { $ref: '#/components/schemas/SourceChannel' }
        source_ref: { type: string }
        customer_id: { $ref: '#/components/schemas/Identifier' }
        customer_name: { type: string }
        receiver_name: { type: string }
        order_status: { $ref: '#/components/schemas/OrderStatus' }
        processing_stage: { $ref: '#/components/schemas/ProcessingStage' }
        processing_health: { $ref: '#/components/schemas/ProcessingHealth' }
        completed_count: { type: integer, minimum: 0 }
        total_count: { type: integer, minimum: 1 }
        attention_reason: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        version: { type: integer, format: int64, minimum: 0 }
    OrderDetail:
      allOf:
        - $ref: '#/components/schemas/OrderSummary'
        - type: object
          required: [receiver, settlement, lines, review_cases]
          properties:
            receiver: { $ref: '#/components/schemas/Receiver' }
            settlement: { $ref: '#/components/schemas/Settlement' }
            remark: { type: string }
            lines:
              type: array
              items: { $ref: '#/components/schemas/OrderLine' }
            review_cases:
              type: array
              items: { $ref: '#/components/schemas/ReviewCase' }
...
    ReviewCase:
      type: object
      required: [id, case_no, case_type, responsible_team, reason_code, status, detail, allowed_actions, version, created_at]
      properties:
        id: { $ref: '#/components/schemas/Identifier' }
        case_no: { type: string }
        case_type: { type: string, minLength: 1, maxLength: 64 }
        responsible_team: { type: string }
        reason_code: { type: string }
        status: { type: string, enum: [OPEN, RESOLVED, DISMISSED] }
        order_id: { $ref: '#/components/schemas/Identifier' }
        order_line_id: { $ref: '#/components/schemas/Identifier' }
        subject_type: { type: string }
        subject_id: { $ref: '#/components/schemas/Identifier' }
        detail: { type: object, additionalProperties: true }
        suggestions:
          type: array
          items: { type: object, additionalProperties: true }
        allowed_actions:
          type: array
          items:
            type: string
            enum: [RESOLVE_CUSTOMER, RESOLVE_SKU, RESOLVE_CARRIER, RESOLVE_IMPORT_DATA, COMPLETE_SOURCE_FOLLOWUP, CONFIRM_ORDER_DRAFT, REJECT_ORDER_DRAFT, CONFIRM_TRACKING_DRAFT, OPEN_SKU_MAPPING, RERUN_JD_SKU_MAPPING_CHECK, RERUN_JD_STOCK_CHECK, REINTERPRET, REJECT]
        resolution: { type: object, additionalProperties: true, nullable: true }
        resolved_by: { type: string, nullable: true }
        resolved_at: { type: string, format: date-time, nullable: true }
        version: { type: integer, minimum: 0 }
        created_at: { type: string, format: date-time }
    ReviewCasePage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/ReviewCase' }
    ResolveCustomerReviewCommand:
      type: object
      required: [expected_version, customer_id, source_channel, source_customer_ref, remark]
      properties:
        expected_version: { type: integer, minimum: 0 }
        customer_id: { $ref: '#/components/schemas/Identifier' }
        source_channel: { $ref: '#/components/schemas/SourceChannel' }
        source_customer_ref: { type: string, minLength: 1 }
        remark: { type: string, minLength: 1, maxLength: 1000 }
    ResolveSkuReviewCommand:
      type: object
      required: [expected_version, sku_id, source_channel, source_sku_ref, quantity_multiplier, remark]
      properties:
        expected_version: { type: integer, minimum: 0 }
        sku_id: { $ref: '#/components/schemas/Identifier' }
        source_channel: { $ref: '#/components/schemas/SourceChannel' }
        source_sku_ref: { type: string, minLength: 1 }
        quantity_multiplier: { $ref: '#/components/schemas/PositiveDecimalQuantity' }
        remark: { type: string, minLength: 1, maxLength: 1000 }
...
    CompleteSourceFollowupCommand:
      type: object
      required: [expected_version, note]
      properties:
        expected_version: { type: integer, minimum: 0 }
        note: { type: string, minLength: 1, maxLength: 1000 }
    VersionedNoteCommand:
      type: object
      required: [expected_version, note]
      properties:
        expected_version: { type: integer, minimum: 0 }
        note: { type: string, minLength: 1, maxLength: 1000 }
...
    ProcurementTicket:
      type: object
      required: [id, ticket_no, fulfillment_id, status, requested_quantity, fulfilled_quantity, remaining_quantity, items, receipts, version, created_at]
      properties:
        id: { $ref: '#/components/schemas/Identifier' }
        ticket_no: { type: string }
        fulfillment_id: { $ref: '#/components/schemas/Identifier' }
        retry_of_ticket_id: { $ref: '#/components/schemas/Identifier' }
        status: { $ref: '#/components/schemas/ProcurementStatus' }
        requested_quantity: { $ref: '#/components/schemas/PositiveDecimalQuantity' }
        fulfilled_quantity: { $ref: '#/components/schemas/DecimalQuantity' }
        remaining_quantity: { $ref: '#/components/schemas/DecimalQuantity' }
        items:
          type: array
          items:
            type: object
            required: [id, sku_id, requested_quantity, fulfilled_quantity, remaining_quantity]
            properties:
              id: { $ref: '#/components/schemas/Identifier' }
              sku_id: { $ref: '#/components/schemas/Identifier' }
              component_sku_id: { $ref: '#/components/schemas/Identifier' }
              requested_quantity: { $ref: '#/components/schemas/PositiveDecimalQuantity' }
              fulfilled_quantity: { $ref: '#/components/schemas/DecimalQuantity' }
              remaining_quantity: { $ref: '#/components/schemas/DecimalQuantity' }
        receipts:
          type: array
          items: { $ref: '#/components/schemas/ProcurementReceipt' }
        version: { type: integer, minimum: 0 }
        created_at: { type: string, format: date-time }
    ProcurementTicketPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/ProcurementTicket' }
    CancelRemainingCommand:
      type: object
      required: [expected_version, reason]
      properties:
        expected_version: { type: integer, minimum: 0 }
        reason: { type: string, minLength: 1, maxLength: 1000 }
...